+++ app/app.js
@@ -2,7 +2,7 @@
 module.exports = function app () { 
 
 const fs = require("fs");
-const startDownloading = require("./startDownloading");
+const startDownloading = require("../startDownloading");
 
 //TODO: Globals aren't ideal but refactoring for sanity to first.
 global.index = -1;
