+++ start.js
@@ -1,5 +1,5 @@
 const prompt = require('prompt');
-const app = require('./app');
+const app = require('./app/app');
 
 const properties = [
   {
@@ -9,6 +9,15 @@ const properties = [
   },
 ];
 
+/* test data 
+uncomment for to enable
+*/
+
+// global.url = "https://music.apple.com/us/album/awakening/350512056"; app(global.url);
+
+
+
+// comment below to disable prompt
 prompt.start();
 
 prompt.get(properties, function (err, result) {
